Remove legacy user indexing
