+++ monitoring/grafana/dashboards/joaquin-playground.json
@@ -1476,96 +1476,6 @@
             ],
             "title": "Average Track State Runtimes",
             "type": "timeseries"
-          },
-          {
-            "fieldConfig": {
-              "defaults": {
-                "color": {
-                  "mode": "palette-classic"
-                },
-                "custom": {
-                  "axisLabel": "Runtime",
-                  "axisPlacement": "auto",
-                  "barAlignment": 0,
-                  "drawStyle": "line",
-                  "fillOpacity": 0,
-                  "gradientMode": "none",
-                  "hideFrom": {
-                    "legend": false,
-                    "tooltip": false,
-                    "viz": false
-                  },
-                  "lineInterpolation": "linear",
-                  "lineWidth": 1,
-                  "pointSize": 5,
-                  "scaleDistribution": {
-                    "type": "linear"
-                  },
-                  "showPoints": "never",
-                  "spanNulls": true,
-                  "stacking": {
-                    "group": "A",
-                    "mode": "none"
-                  },
-                  "thresholdsStyle": {
-                    "mode": "off"
-                  }
-                },
-                "mappings": [],
-                "thresholds": {
-                  "mode": "absolute",
-                  "steps": [
-                    {
-                      "color": "green"
-                    },
-                    {
-                      "color": "red",
-                      "value": 80
-                    }
-                  ]
-                },
-                "unit": "s"
-              },
-              "overrides": []
-            },
-            "gridPos": {
-              "h": 8,
-              "w": 12,
-              "x": 12,
-              "y": 14
-            },
-            "id": 26,
-            "options": {
-              "legend": {
-                "calcs": [
-                  "mean",
-                  "lastNotNull",
-                  "max"
-                ],
-                "displayMode": "table",
-                "placement": "bottom",
-                "sortBy": "Max",
-                "sortDesc": true
-              },
-              "tooltip": {
-                "mode": "multi",
-                "sort": "desc"
-              }
-            },
-            "targets": [
-              {
-                "datasource": {
-                  "type": "prometheus"
-                },
-                "exemplar": true,
-                "expr": "max by (scope) (rate(audius_dn_user_state_update_duration_seconds_sum{environment=~\"$env\", host=~\"$host\"}[5m]) / rate(audius_dn_user_state_update_duration_seconds_count{environment=~\"$env\", host=~\"$host\"}[5m]))",
-                "interval": "",
-                "legendFormat": "{{host}}{{scope}}",
-                "refId": "A"
-              }
-            ],
-            "title": "Average User State Runtimes",
-            "type": "timeseries"
           }
         ],
         "title": "State Metrics",
